Finale Version Entlassen/wiedereinstellen User-Profil anschauen Passwort zurücksetzen

+++ src/Controller/PlanesController.php
@@ -2,6 +2,10 @@
 namespace App\Controller;
 
 use App\Controller\AppController;
+use Cake\Event\Event;
+use Cake\Network\Exception\NotFoundException;
+use Cake\Network\Exception\BadRequestException;
+use Cake\Network\Exception\UnauthorizedException;
 
 /**
  * Planes Controller
@@ -10,6 +14,19 @@ use App\Controller\AppController;
  */
 class PlanesController extends AppController
 {
+
+    public function beforeFilter(Event $event){
+        parent::beforeFilter($event);
+
+        if( $this->Auth->user() ){ //eingelogget
+            if( (in_array($this->Auth->user('group_id'), ['4']) && in_array($this->request->action, ['index', 'view', 'add', 'edit', 'delete'])) ||
+                (in_array($this->Auth->user('group_id'), ['1','2','3']) && in_array($this->request->action, ['index', 'view'])) ){ // admin
+                // ok
+            }else{
+                throw new UnauthorizedException();
+            }
+        }
+    }
 
     /**
      * Index method
@@ -100,10 +117,14 @@ class PlanesController extends AppController
     {
         $this->request->allowMethod(['post', 'delete']);
         $plane = $this->Planes->get($id);
-        if ($this->Planes->delete($plane)) {
-            $this->Flash->success('Das Flugzeug wurde glöscht..');
-        } else {
-            $this->Flash->error('Das zu löschende Flugzeug wurde nicht gefunden. Bitte erneut versuchen.');
+        if(!$this->Planes->Flights->exists(['plane_id' => $plane->id,'status IN' =>[FLIGHT_SOON, FLIGHT_FLYING]])){
+            if ($this->Planes->delete($plane)) {
+                $this->Flash->success('Das Flugzeug wurde glöscht..');
+            } else {
+                $this->Flash->error('Das zu löschende Flugzeug wurde nicht gefunden. Bitte erneut versuchen.');
+            }
+        }else{
+            $this->Flash->error('Das Flugzeug ist noch verbucht.');
         }
         return $this->redirect(['action' => 'index']);
     }
